fixed network

+++ Aurea/Assets/01_Scripts/Fight/FightController.cs
@@ -42,15 +42,15 @@ public class FightController : MonoBehaviourPunCallbacks
     [SerializeField]
     public List<PlayerController> players = new List<PlayerController>();
 
-    [SerializeField]
-    private PlayerController oldplayer = null;
+    // [SerializeField]
+    // private PlayerController oldplayer = null;
     // [SerializeField]
     // private RandomPlayerController randomPlayer = null;
 
-    [SerializeField]
-    private PlayerController olsenemy = null;
-    [SerializeField]
-    private AgentController enemyAgent = null;
+    // [SerializeField]
+    // private PlayerController olsenemy = null;
+    // [SerializeField]
+    // private AgentController enemyAgent = null;
 
     [SerializeField]
     private List<GameObject> playerSpawnpoints = new List<GameObject>();
@@ -65,6 +65,10 @@ public class FightController : MonoBehaviourPunCallbacks
     private List<GameObject> enemyCrystals = new List<GameObject>();
     [SerializeField]
     private List<Transform> enemyCamPositions = new List<Transform>();
+    [SerializeField]
+    private GameObject playAgainButton = null;
+    [SerializeField]
+    private GameObject returnButton = null;
 
     [SerializeField]
     public float roundTime = 30f;
@@ -253,10 +257,10 @@ public class FightController : MonoBehaviourPunCallbacks
 
     void LoadData()
     {
-        oldplayer.SetData(Player.Instance);
-        LoadedPlayer?.Invoke(oldplayer);
+        // oldplayer.SetData(Player.Instance);
+        // LoadedPlayer?.Invoke(oldplayer);
 
-        LoadedEnemy?.Invoke(oldplayer);
+        // LoadedEnemy?.Invoke(oldplayer);
     }
 
     int FlipCoin()
@@ -285,23 +289,23 @@ public class FightController : MonoBehaviourPunCallbacks
         PlayerController lostPlayer = players[0].view.ViewID == _playerID ? players[0] : players[1];
 
         // PlayerController losePlayer = PhotonView.Find(_playerID).GetComponent<PlayerController>();
+        // NetworkController.instance.Kill();
         StartCoroutine(CloseIn(3f));
+
+        winPlayer.Won();
+
         if (lostPlayer.view.Owner.IsLocal)
         {
             gameOverText.text = "You lose!";
-            winPlayer.Won();
-            // SceneManager.LoadScene("Lose");
-            //     data.AddLoseStatistics();
-            //     enemyData.AddWonStatistics();
+            // SceneManager.LoadSceneAsync("Lose");
         }
         else
         {
             gameOverText.text = "You won!";
-            winPlayer.Won();
-            // SceneManager.LoadScene("Won");
+            // SceneManager.LoadSceneAsync("Won");
         }
 
-        
+
 
         gameOverScreen.SetActive(true);
 
@@ -383,10 +387,10 @@ public class FightController : MonoBehaviourPunCallbacks
 
     public PlayerController GetPlayer()
     {
-        return oldplayer;
+        return players[0];
     }
 
-    public PlayerController GetEnemy() { return olsenemy; }
+    public PlayerController GetEnemy() { return players[0]; }
 
     public AureaData GetAureaData(string _name)
     {
@@ -399,9 +403,18 @@ public class FightController : MonoBehaviourPunCallbacks
         return null;
     }
 
-    IEnumerator CloseIn(float _time) {
+    IEnumerator CloseIn(float _time)
+    {
+        playAgainButton.SetActive(false);
+        returnButton.SetActive(false);
+        
         yield return new WaitForSeconds(_time);
-        NetworkController.instance.Kill();
+
+        if (NetworkController.instance)
+            NetworkController.instance.Kill();
+
+        playAgainButton.SetActive(true);
+        returnButton.SetActive(true);
     }
     IEnumerator WaitBetweenClick()
     {
